More random stuff added. Loading/Saving implemented, but still needs to be added into game flow

+++ Asteroids/src/views/Bullet.java
@@ -6,7 +6,7 @@ import java.awt.geom.GeneralPath;
 import models.Updatable;
 
 public class Bullet extends BaseView implements Updatable {
-	public Bullet(int x, int y, double angle, int vx, int vy, int vAngle, Color bulletColor){
+	public Bullet(int x, int y, double angle, double vx, double vy, double vAngle, Color bulletColor){
 		model = new BulletModel(x,y,angle,vx,vy,vAngle);
 		color = bulletColor;
 		shape = new GeneralPath(4);
